+++ packages/spright-components/src/chat/message/index.ts
@@ -34,7 +34,7 @@ export class ChatMessage extends FoundationElement {
      * HTML Attribute: message-type
      */
     @attr({ attribute: 'message-type' })
-    public readonly messageType: ChatMessageType = ChatMessageType.system;
+    public messageType: ChatMessageType = ChatMessageType.system;
 
     /** @internal */
     @observable