+++ packages/blazor-workspace/SprightBlazor/CHANGELOG.md
@@ -1,8 +1,32 @@
 # Change Log - @ni/spright-blazor
 
-<!-- This log was last generated on Fri, 20 Jun 2025 06:07:44 GMT and should not be manually modified. -->
+<!-- This log was last generated on Fri, 27 Jun 2025 15:07:42 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 4.3.4
+
+Fri, 27 Jun 2025 15:07:42 GMT
+
+### Patches
+
+- Bump @ni/spright-components to v5.5.0
+
+## 4.3.3
+
+Thu, 26 Jun 2025 19:20:55 GMT
+
+### Patches
+
+- Bump @ni/spright-components to v5.4.3
+
+## 4.3.2
+
+Tue, 24 Jun 2025 16:53:02 GMT
+
+### Patches
+
+- Bump @ni/spright-components to v5.4.2
 
 ## 4.3.1